python_server/requirements.txt
Flask
playwright
python-dotenv
gunicorn